third_party/WebKit/LayoutTests/svg/text/getextentofchar-nonbmp.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<svg>
  <text font-size="50" y="50">&#x1D11E;&#x1D11E;</text>
</svg>
<script>
test(function() {
  var text = document.querySelector('text');
  assert_equals(text.getNumberOfChars(), 4);
  assert_equals(text.getExtentOfChar(0).width, text.getExtentOfChar(1).width);
  assert_equals(text.getExtentOfChar(2).width, text.getExtentOfChar(3).width);
}, 'SVGTextContentElement.getExtentOfChar, non-BMP characters.');
</script>
